created class for messages, moved it with UserRole to new repository
