+++ src/main/java/com/bookshop/repository/user/UserRoleRepository.java
@@ -1,6 +1,6 @@
 package com.bookshop.repository.user;
 
-import com.bookshop.model.UserRole;
+import com.bookshop.res.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRoleRepository extends JpaRepository<UserRole, Long> {
